Don't go to dashboardRoute until auth is completed
Removed extra routing to dashboardRoute since it was causing the app to redirect to dashboardRoute before the callback was finished (and therefore before the user had even authenticated! Major security flaw to say the least). The callback already handles routing correctly if there is no error, so there was no need for the extra routing.